Move SummaryPresenter into its own folder
We renamed SummaryPresenter to EscortPresenter and
moved the logic into a folder to be consistent with
the work done with the PDF generation.
