+++ app/views/escorts/summary.html.erb
@@ -7,7 +7,7 @@
     <header>
       <h1><%= t(".heading") %></h1>
     </header>
-    <%= render('escorts/summary/prisoner', prisoner: @summary.prisoner) %>
-    <%= render('escorts/summary/move', move: @summary.move) %>
+    <%= render('escorts/summary/prisoner', prisoner: @escort.prisoner) %>
+    <%= render('escorts/summary/move', move: @escort.move) %>
   </div>
 </div>
